+++ src/states/entity/BaseEntity.lua
@@ -7,6 +7,11 @@ function BaseEntity:init(params)
 end
 
 function BaseEntity:render()
+    if self.type == 'Cursor' then
+        self.x = mouseX
+        self.y = mouseY
+    end
+
     love.graphics.draw(self.frame, self.x, self.y, 0, 
         self.desired_width / self.frame:getWidth(), self.desired_height / self.frame:getHeight())
 end
@@ -14,12 +19,4 @@ end
 function BaseEntity:isMouseOver()
     return mouseX > self.x and mouseX < self.x + self.desired_width and
            mouseY > self.y and mouseY < self.y + self.desired_height
-end
-
-function BaseEntity:onPressed()
-    if self.isMachine and self.productionStage == 'Void' then
-        self:produce()
-    elseif self.isGUI then
-        self:clicked()
-    end
 end